+++ lint.sh
@@ -2,6 +2,10 @@
 
 npm install
 
-./node_modules/.bin/eslint ramls/loan-storage.raml
+./node_modules/.bin/raml-cop ramls/loan-storage.raml
 
-./node_modules/.bin/eslint ramls/loan-policy-storage.raml
+./node_modules/.bin/raml-cop ramls/loan-policy-storage.raml
+
+./node_modules/.bin/raml-cop ramls/loan-rules-storage.raml
+
+./node_modules/.bin/raml-cop ramls/request-storage.raml